requirements.txt
# ============================================================================
# Aviation Chatbot - Production Requirements
# Python 3.11+
# ============================================================================

# -------------------------
# Web UI
# -------------------------
streamlit==1.36.0

# -------------------------
# Database & Vector Search
# -------------------------
psycopg2-binary==2.9.11
sqlalchemy==2.0.45
pgvector==0.4.2

# -------------------------
# Embeddings
# -------------------------
sentence-transformers==2.7.0
torch==2.2.0
transformers==4.57.6
huggingface-hub==0.36.0

# -------------------------
# LLM (Cloud)
# -------------------------
google-generativeai==0.8.6

# -------------------------
# PDF Processing
# -------------------------
pdfplumber==0.11.0
pypdf==4.2.0

# -------------------------
# Text Chunking
# -------------------------
langchain==0.2.11
langchain-text-splitters==0.2.4

# -------------------------
# Data Handling
# -------------------------
numpy==1.26.4
pandas==2.2.2

# -------------------------
# Utilities
# -------------------------
python-dotenv==1.2.1
tqdm==4.67.1

# BM25 for hybrid search (sparse retrieval)
rank-bm25>=0.2.2

# Cross-encoder for re-ranking (from sentence-transformers)
# Already installed with sentence-transformers, but listing explicitly
# sentence-transformers includes cross-encoder models

# Optional: For faster BM25 implementation
python-Levenshtein>=0.12.0  # Uncomment if you want fuzzy matching

# ============================================================================
# OPTIONAL (for development/testing)
# ============================================================================
# jupyter==1.0.0
# ipykernel==6.29.0
# ipywidgets==8.1.1